task(#44): Clean up
- make login and signup shared code into a reusable scrollable form container
- fixes to bugs in map routes
