+++ apps/driver-app/app/login.tsx
@@ -1,34 +1,10 @@
-import KeyboardDismiss from "@/src/components/keyboard-dismiss";
-import LoginForm from "@/src/components/login-form";
-import { KeyboardAvoidingView, Platform, StyleSheet, View } from "react-native";
-import { ScrollView } from "react-native-gesture-handler";
-import { SafeAreaView } from "react-native-safe-area-context";
+import LoginForm from '@/src/components/login-form';
+import ScrollableFormContainer from '@/src/components/scrollable-form-container';
 
 export default function Login() {
   return (
-    <SafeAreaView style={{ flex: 1 }}>
-      <KeyboardAvoidingView
-        behavior={Platform.OS === "ios" ? "padding" : "height"}
-        style={{ flex: 1 }}
-      >
-        <ScrollView
-          contentContainerStyle={{
-            flexGrow: 1,
-          }}
-        >
-          <KeyboardDismiss>
-            <View style={styles.container}>
-              <LoginForm />
-            </View>
-          </KeyboardDismiss>
-        </ScrollView>
-      </KeyboardAvoidingView>
-    </SafeAreaView>
+    <ScrollableFormContainer>
+      <LoginForm />
+    </ScrollableFormContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
