+++ apps/driver-app/app/signup.tsx
@@ -1,35 +1,10 @@
-import KeyboardDismiss from "@/src/components/keyboard-dismiss";
-import SignUpForm from "@/src/components/sign-up-form";
-import React from "react";
-import { KeyboardAvoidingView, Platform, StyleSheet, View } from "react-native";
-import { ScrollView } from "react-native-gesture-handler";
-import { SafeAreaView } from "react-native-safe-area-context";
+import ScrollableFormContainer from '@/src/components/scrollable-form-container';
+import SignUpForm from '@/src/components/sign-up-form';
 
 export default function SignUp() {
   return (
-    <SafeAreaView style={{ flex: 1 }}>
-      <KeyboardAvoidingView
-        behavior={Platform.OS === "ios" ? "padding" : "height"}
-        style={{ flex: 1 }}
-      >
-        <ScrollView
-          contentContainerStyle={{
-            flexGrow: 1,
-          }}
-        >
-          <KeyboardDismiss>
-            <View style={styles.container}>
-              <SignUpForm />
-            </View>
-          </KeyboardDismiss>
-        </ScrollView>
-      </KeyboardAvoidingView>
-    </SafeAreaView>
+    <ScrollableFormContainer>
+      <SignUpForm />
+    </ScrollableFormContainer>
   );
 }
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-  },
-});
