+++ apps/driver-app/src/components/scrollable-form-container/index.tsx
@@ -1,0 +1,33 @@
+import KeyboardDismiss from '@/src/components/keyboard-dismiss';
+import { KeyboardAvoidingView, Platform, StyleSheet, View } from 'react-native';
+import { ScrollView } from 'react-native-gesture-handler';
+import { SafeAreaView } from 'react-native-safe-area-context';
+
+const ScrollableFormContainer = ({ children }) => {
+  return (
+    <SafeAreaView style={{ flex: 1 }}>
+      <KeyboardAvoidingView
+        behavior={Platform.OS === 'ios' ? 'padding' : 'height'}
+        style={{ flex: 1 }}
+      >
+        <ScrollView
+          contentContainerStyle={{
+            flexGrow: 1
+          }}
+        >
+          <KeyboardDismiss>
+            <View style={styles.container}>{children}</View>
+          </KeyboardDismiss>
+        </ScrollView>
+      </KeyboardAvoidingView>
+    </SafeAreaView>
+  );
+};
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1
+  }
+});
+
+export default ScrollableFormContainer;
