+++ apps/driver-app/src/components/setup.tsx
@@ -10,7 +10,7 @@ export const Setup = () => {
     if (!authToken) {
       router.replace('/signup');
     } else {
-      router.replace('/map');
+      router.replace('/main');
     }
   }, []);
 
@@ -18,7 +18,7 @@ export const Setup = () => {
     <Stack initialRouteName="signup">
       <Stack.Screen name="signup" options={{ headerShown: false }} />
       <Stack.Screen name="login" options={{ headerShown: false }} />
-      <Stack.Screen name="map" options={{ headerShown: false }} />
+      <Stack.Screen name="main" options={{ headerShown: false }} />
     </Stack>
   );
 };
